requirements.txt
django
gunicorn
django-heroku
django-crispy-forms
dj-database-url
pytz
